core/configs: Rename cell flag "Unmanaged Exit" to "No Comm Region"
The purpose of this flag is to enable simple cells that do not
participate in the Communication Region protocols. The fact that we can
be shut down without prior consultation is one result of this, but will
not be the only one in the future. So rename the flag to better reflect
its semantic.

Note that we will continue to initialize the communication region of a
cell even if the flag is set for it. It is simply not worth to consider
this case and skip the code related code.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>