core: Add Cell Reload hypercall
Provide the new hypercall "Cell Reload". It has the semantic of
suspending an existing cell and remapping all its memory regions that
are marked reloadable back to the root cell.  On cell start, all its
reloadable regions that are mapped to the root cell will be unmapped
from it again.

This enables reloading of non-root cells without the so far required
cell destruction/recreation cycle.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>